requirements.txt
streamlit
pandas
jinja2
weasyprint
streamlit-drawable-canvas
Pillow
numpy
pandas
plotly
openpyxl
requests
pymupdf
streamlit-image-coordinates